refactor: useCoupon 및 useProduct 훅을 features 디렉토리로 이동
